+++ Microsoft.OpenAIRateLimiter.Service/ITokenService.cs
@@ -2,7 +2,7 @@
 {
     public interface ITokenService
     {
-        double CalculateCost(int TotalTokens, string model);
+        decimal CalculateCost(int TotalTokens, string model);
         Task<int> GetTokenCount(string prompt, string model);
     }
 }
